refactor: made code reorder and compact
